Move /stores and Notification.jsx from ui/job-view into ui/shared
