+++ ui/shared/stores/selectedJobStore.js
